Validate a data frame with predator/prey mass ratio observations

The data frame needs to have the columns "species", "w_pred", "w_prey" and "n_prey". The function adds a log_ppmr column to the data frame with the logarithm of the predator/prey mass ratio.

Usage

validate_ppmr_data(ppmr_data, species = NULL)

Arguments

ppmr_data

A data frame with log ppmr observations

species

The species to select. Optional. If supplied the function only selects the rows for the specified species. If the species is not found, an error is raised.

Value

Valid ppmr data frame with columns "species", "w_pred", "w_prey", "n_prey" and "log_ppmr" and rows for the specified species.

Details

If a log_ppmr column already exists, it is overwritten and a warning is issued if discrepancies were found. A warning is also issued if any prey is heavier than the predator.

See also